Function dokan::is_name_in_expression

pub fn is_name_in_expression(
    expression: impl AsRef<U16CStr>, 
    name: impl AsRef<U16CStr>, 
    ignore_case: bool
) -> bool

Checks whether the name matches the specified expression.

This is a helper function that can be used to implement FileSystemHandler::find_files_with_pattern. It behaves like the FsRtlIsNameInExpression routine provided for file system drivers by Windows.