Struct dokan::Drive

pub struct Drive<'a> { /* fields omitted */ }

A builder that allows configuring and mounting a volume.

Methods

impl<'a> Drive<'a>

pub fn new() -> Self

Creates a new instance of this builder with default settings.

pub fn thread_count(&mut self, value: u16) -> &mut Self

Sets the number of threads used to handle file system operations.

pub fn flags(&mut self, value: MountFlags) -> &mut Self

Sets flags that controls behavior of the volume.

pub fn mount_point(&mut self, value: &'a impl AsRef<U16CStr>) -> &mut Self

Sets mount point path.

pub fn unc_name(&mut self, value: &'a impl AsRef<U16CStr>) -> &mut Self

Sets UNC name of the network drive.

pub fn timeout(&mut self, value: Duration) -> &mut Self

Sets the time that Dokan will wait for an operation to complete.

If an operation times out, the user mode implementation is considered to be unable to handle file system operations properly, and the driver will therefore unmount the volume in order to keep the system stable.

This timeout can be temporarily extended for an operation with OperationInfo::reset_timeout.

pub fn allocation_unit_size(&mut self, value: u32) -> &mut Self

Sets allocation unit size of the volume.

This value will affect file sizes.

pub fn sector_size(&mut self, value: u32) -> &mut Self

Sets sector size of the volume.

This value will affect file sizes.

pub fn mount<'b, 'c: 'b, T: FileSystemHandler<'b, 'c> + 'c>(
&mut self,
handler: &'c T
) -> Result<(), MountError>

Mounts the volume and blocks the current thread until the volume gets unmounted.