Struct dokan::OperationInfo

pub struct OperationInfo<'a, 'b: 'a, T: FileSystemHandler<'a, 'b> + 'b> { /* fields omitted */ }

Information about the current operation.

Implementations

impl<'a, 'b: 'a, 'c: 'b, T: FileSystemHandler<'b, 'c> + 'c> OperationInfo<'b, 'c, T>

pub fn pid(&self) -> u32

Gets process ID of the calling process.

pub fn is_dir(&self) -> bool

Gets whether the target file is a directory.

pub fn delete_on_close(&self) -> bool

Gets whether the file should be deleted when it is closed.

pub fn paging_io(&self) -> bool

Gets whether it is a paging I/O operation.

pub fn synchronous_io(&self) -> bool

Gets whether it is a synchronous I/O operation.

pub fn no_cache(&self) -> bool

Gets whether it is a non-cached I/O operation.

pub fn write_to_eof(&self) -> bool

Gets whether the current write operation should write to end of file instead of the position specified by the offset argument.

pub fn thread_count(&self) -> u16

Gets the number of threads used to handle file system operations.

pub fn mount_flags(&self) -> MountFlags

Gets flags that controls behavior of the mounted volume.

pub fn mount_point(&self) -> Option<&U16CStr>

Gets mount point path.

pub fn unc_name(&self) -> Option<&U16CStr>

Gets UNC name of the network drive.

pub fn timeout(&self) -> Duration

Gets the time that Dokan will wait for an operation to complete.

See Drive::timeout for more information.

pub fn allocation_unit_size(&self) -> u32

Gets allocation unit size of the volume.

pub fn sector_size(&self) -> u32

Gets sector size of the volume.

#[must_use]pub fn reset_timeout(&self, timeout: Duration) -> bool

Temporarily extend the timeout of the current operation.

Returns true on success.

pub fn requester_token(&self) -> Option<TokenHandle>

Gets the access token associated with the calling process.

Returns None on error.